refactor: move http_server.h to dsn/http (#592)
