+++ src/http/pprof_http_service.h
@@ -6,7 +6,7 @@
 
 #ifdef DSN_ENABLE_GPERF
 
-#include <dsn/tool-api/http_server.h>
+#include <dsn/http/http_server.h>
 
 namespace dsn {
