+++ src/meta/meta_backup_service.h
@@ -6,7 +6,7 @@
 #include <functional>
 
 #include <dsn/dist/block_service.h>
-#include <dsn/tool-api/http_server.h>
+#include <dsn/http/http_server.h>
 #include <dsn/perf_counter/perf_counter_wrapper.h>
 
 #include "meta_data.h"
