+++ src/meta/meta_http_service.h
@@ -6,7 +6,7 @@
 
 #include <algorithm>
 
-#include <dsn/tool-api/http_server.h>
+#include <dsn/http/http_server.h>
 
 namespace dsn {
 namespace replication {
